fix flicker on ios mode with android devices
